name_resolve: Split into modules
